added skeleton for the booking screen
